pet/requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.89.1
Flask==2.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pydantic==1.10.4
pytest==7.2.0
python-multipart==0.0.5
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.4.0
unicorn==2.0.1.post1
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.2
zipp==3.11.0
